+++ RELEASE_NOTES.md
@@ -5,6 +5,8 @@
 * Updates dxCommon version, which fixes the output of the `printTree` command
 * Allow primitive -> `String` coercion for WDL versions <= 1.0
 * Fixes parsing of unary operations
+* A `parameter_meta` key that does not correspond to an input or output parameter is now ignored with a warning
+* Fixes caching of imports during parsing - now absolute rather than relative file paths are used 
 
 ## 0.14.1 (2021-06-08)